Move Shape trait into fidget::shape
