+++ fidget/benches/mesh.rs
@@ -1,7 +1,7 @@
 use criterion::{
     black_box, criterion_group, criterion_main, BenchmarkId, Criterion,
 };
-use fidget::eval::MathShape;
+use fidget::shape::MathShape;
 
 const COLONNADE: &str = include_str!("../../models/colonnade.vm");
