+++ fidget/benches/render.rs
@@ -4,7 +4,7 @@ use criterion::{
 
 const PROSPERO: &str = include_str!("../../models/prospero.vm");
 
-use fidget::{eval::MathShape, shape::RenderHints};
+use fidget::shape::{MathShape, RenderHints};
 
 pub fn prospero_size_sweep(c: &mut Criterion) {
     let (ctx, root) = fidget::Context::from_text(PROSPERO.as_bytes()).unwrap();
