+++ fidget/src/core/context/mod.rs
@@ -733,7 +733,7 @@ impl Context {
     /// Evaluates the given node with the provided values for X, Y, and Z.
     ///
     /// This is extremely inefficient; consider converting the node into a
-    /// [`Shape`](crate::eval::Shape) and using its evaluators instead.
+    /// [`Shape`](crate::shape::Shape) and using its evaluators instead.
     ///
     /// ```
     /// # let mut ctx = fidget::context::Context::new();
@@ -762,7 +762,7 @@ impl Context {
     /// Evaluates the given node with a generic set of variables
     ///
     /// This is extremely inefficient; consider converting the node into a
-    /// [`Shape`](crate::eval::Shape) and using its evaluators instead.
+    /// [`Shape`](crate::shape::Shape) and using its evaluators instead.
     pub fn eval(
         &self,
         root: Node,
