+++ fidget/src/core/eval/test/grad_slice.rs
@@ -5,7 +5,7 @@
 use super::{build_stress_fn, test_args, CanonicalBinaryOp, CanonicalUnaryOp};
 use crate::{
     context::Context,
-    eval::{BulkEvaluator, EzShape, MathShape, Shape},
+    shape::{BulkEvaluator, EzShape, MathShape, Shape},
     types::Grad,
 };
