+++ fidget/src/core/shape/bulk.rs
