+++ fidget/src/core/shape/tracing.rs
@@ -12,8 +12,8 @@ use crate::{eval::Tape, Error};
 
 /// Evaluator for single values which simultaneously captures an execution trace
 ///
-/// The trace can later be used to simplify the [`Shape`](crate::eval::Shape)
-/// using [`Shape::simplify`](crate::eval::Shape::simplify).
+/// The trace can later be used to simplify the [`Shape`](crate::shape::Shape)
+/// using [`Shape::simplify`](crate::shape::Shape::simplify).
 pub trait TracingEvaluator: Default {
     /// Data type used during evaluation
     type Data: From<f32> + Copy + Clone;
